pyproject.toml
[project]
name = "relearn-python"
version = "1.0.0"
description = "Python Mastery - Apprentissage complet de Python"
readme = "README.md"
requires-python = ">= 3.13"
authors = [
    { name = "Ranker", email = "ranker@example.com" }
]
dependencies = [
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "tabulate>=0.9.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
core = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
]
web = [
    "flask>=3.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "jinja2>=3.1.0",
]
automation = [
    "beautifulsoup4>=4.12.0",
    "selenium>=4.10.0",
    "webdriver-manager>=4.0.0",
]
data = [
    "scikit-learn>=1.3.0",
    "openpyxl>=3.1.0",
    "pillow>=10.0.0",
]
ai = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]

[tool.ruff]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
